build error 수정2

+++ src/main/java/com/example/BE/review/ReviewRepository.java
@@ -63,5 +63,5 @@ public interface ReviewRepository extends JpaRepository<ReviewEntity, Integer> {
     @Query("SELECT r FROM ReviewEntity r WHERE r.user.userId = :userId")
     List<ReviewEntity> findPosterByUserId(@Param("userId") int userId);
 
-    List<ReviewEntity> findByUser_UserIdAndRatingGreaterThanEqual(Long userId);
+    List<ReviewEntity> findByUser_UserIdAndRatingGreaterThanEqual(int userId);
 }
